修改点菜js、后台bill的fee字段属性改为float

+++ src/main/java/com/fandian/bean/Bill.java
@@ -15,7 +15,7 @@ public class Bill {
 
     private int status;
 
-    private int fee;
+    private float fee;
 
     private float actualFee;
 
@@ -113,11 +113,11 @@ public class Bill {
         this.billDetails = billDetails;
     }
 
-    public int getFee() {
+    public float getFee() {
         return fee;
     }
 
-    public void setFee(int fee) {
+    public void setFee(float fee) {
         this.fee = fee;
     }
